[CSL-1800] Move filesystem utlities into Pos.Util.Filesystem (cardano-sl-util)
